+++ lib/linked_map/duplicate_value_error.ex
@@ -1,0 +1,7 @@
+defmodule LinkedMap.DuplicateValueError do
+  defexception [:value, :message]
+
+  @impl true
+  def message(%{value: value, message: nil}), do: "value #{inspect(value)} is already present"
+  def message(%{message: message}), do: message
+end