+++ lib/linked_map/missing_value_error.ex
@@ -1,0 +1,7 @@
+defmodule LinkedMap.MissingValueError do
+  defexception [:value, :message]
+
+  @impl true
+  def message(%{value: value, message: nil}), do: "value #{inspect(value)} is not present"
+  def message(%{message: message}), do: message
+end