+++ test/linked_map_test.exs
@@ -3,11 +3,11 @@ defmodule LinkedMapTest do
   alias LinkedMap.Node
   doctest LinkedMap
 
-  test "creating a new map" do
+  test "new/0" do
     assert LinkedMap.new() == %LinkedMap{head: nil, tail: nil, items: %{}}
   end
 
-  test "adding to an empty map" do
+  test "add/2 with an empty map" do
     result = LinkedMap.new() |> LinkedMap.add("a")
 
     assert result.head == "a"
@@ -17,7 +17,7 @@ defmodule LinkedMapTest do
     assert result.items["a"].next == nil
   end
 
-  test "adding to a single item map" do
+  test "add/2 with a single item map" do
     map = LinkedMap.new() |> LinkedMap.add("a")
     result = map |> LinkedMap.add("b")
 
@@ -30,7 +30,7 @@ defmodule LinkedMapTest do
     assert result.items["b"].next == nil
   end
 
-  test "adding to a N item map" do
+  test "add/2 with a N item map" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b")
     result = map |> LinkedMap.add("c")
 
@@ -45,28 +45,58 @@ defmodule LinkedMapTest do
     assert result.items["c"].next == nil
   end
 
-  test "removing from an empty map" do
+  test "add_new/2 with a new item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    result = map |> LinkedMap.add_new("b")
+
+    assert Map.keys(result.items) == ["a", "b"]
+  end
+
+  test "add_new/2 with an existing item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    result = map |> LinkedMap.add_new("a")
+
+    assert Map.keys(result.items) == ["a"]
+  end
+
+  test "add_new!/2 with a new item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    result = map |> LinkedMap.add_new!("b")
+
+    assert Map.keys(result.items) == ["a", "b"]
+  end
+
+  test "add_new!/2 with an existing item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    message = ~s(value "a" is already present)
+
+    assert_raise LinkedMap.DuplicateValueError, message, fn ->
+      LinkedMap.add_new!(map, "a")
+    end
+  end
+
+  test "remove/2 with an empty map" do
     map = LinkedMap.new()
     result = map |> LinkedMap.remove("foo")
 
     assert result == map
   end
 
-  test "removing a non-existent item" do
+  test "remove/2 with a non-existent item" do
     map = LinkedMap.new() |> LinkedMap.add("a")
     result = map |> LinkedMap.remove("foo")
 
     assert result == map
   end
 
-  test "removing the only item" do
+  test "remove/2 with the only item" do
     map = LinkedMap.new() |> LinkedMap.add("a")
     result = map |> LinkedMap.remove("a")
 
     assert result == LinkedMap.new()
   end
 
-  test "removing the first of two items" do
+  test "remove/2 with the first of two items" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b")
     result = map |> LinkedMap.remove("a")
 
@@ -77,7 +107,7 @@ defmodule LinkedMapTest do
     assert result.items["b"].next == nil
   end
 
-  test "removing the last of two items" do
+  test "remove/2 with the last of two items" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b")
     result = map |> LinkedMap.remove("b")
 
@@ -88,7 +118,7 @@ defmodule LinkedMapTest do
     assert result.items["a"].next == nil
   end
 
-  test "removing the first of N items" do
+  test "remove/2 with the first of N items" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b") |> LinkedMap.add("c")
     result = map |> LinkedMap.remove("a")
 
@@ -101,7 +131,7 @@ defmodule LinkedMapTest do
     assert result.items["c"].next == nil
   end
 
-  test "removing the last of N items" do
+  test "remove/2 with the last of N items" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b") |> LinkedMap.add("c")
     result = map |> LinkedMap.remove("c")
 
@@ -114,7 +144,7 @@ defmodule LinkedMapTest do
     assert result.items["b"].next == nil
   end
 
-  test "removing a random of N items" do
+  test "remove/2 from the middle of N items" do
     map = LinkedMap.new() |> LinkedMap.add("a") |> LinkedMap.add("b") |> LinkedMap.add("c")
     result = map |> LinkedMap.remove("b")
 
@@ -125,5 +155,21 @@ defmodule LinkedMapTest do
     assert result.items["a"].next == "c"
     assert result.items["c"].previous == "a"
     assert result.items["c"].next == nil
+  end
+
+  test "remove!/2 with an existing item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    result = map |> LinkedMap.remove!("a")
+
+    assert Map.keys(result.items) == []
+  end
+
+  test "remove!/2 with a non-existant item" do
+    map = LinkedMap.new() |> LinkedMap.add("a")
+    message = ~s(value "b" is not present)
+
+    assert_raise LinkedMap.MissingValueError, message, fn ->
+      LinkedMap.remove!(map, "b")
+    end
   end
 end