+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v0.2.0 (pending)
+
+### New Features
+
+- Added `add_new/2` which only adds an item if it is not already present
+- Added `add_new!/2` which adds an item if it is not already present, or raises if it is
+- Added `remove!/2` which removes an item if it exists, or raises if it doesn't
+
+### Bug Fixes
+
+- Fix some incorrect references to `__MODULE__` in docs
+
+---
+
 ## v0.1.0
 
 Initial release